Split out tilemap shuffler in US to make heart thing easier
