+++ config/splat.us.stnp3.yaml
@@ -165,7 +165,8 @@ segments:
       - [0x32830, c, entrance_stage_entities]
       - [0x32F30, c, e_castle_door]
       - [0x332A8, c, entrance_stage_entities_2]
-      - [0x363FC, c, e_explosion_puff_opaque]
+      - [0x3644C, c, e_tilemap_shuffler]
+      - [0x3653C, c, e_explosion_puff_opaque]
       - [0x36990, c, slogra]
       - [0x378BC, c, gaibon]
       - [0x390BC, c, st_update]
