+++ src/st/np3/e_tilemap_shuffler.c
@@ -1,0 +1,4 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "np3.h"
+
+#include "../e_tilemap_shuffler.h"
